refactor: move transaction execution out of blockstore_processor to runtime
