+++ ledger/src/lib.rs
@@ -32,7 +32,6 @@ pub mod sigverify_shreds;
 pub mod slot_stats;
 mod staking_utils;
 mod transaction_address_lookup_table_scanner;
-pub mod transaction_balances;
 pub mod use_snapshot_archives_at_startup;
 
 #[macro_use]
