+++ runtime/src/lib.rs
@@ -50,7 +50,9 @@ pub mod stakes;
 pub mod static_ids;
 pub mod status_cache;
 pub mod test_utils;
+pub mod transaction_balances;
 pub mod transaction_batch;
+pub mod transaction_execution;
 pub mod validated_block_finalization;
 pub mod validated_reward_certificate;
 pub mod vote_sender_types;
